Split views into separate modules for API and UI
